optimize component spec
