+++ pkg/manager/member/pvc_resizer.go
@@ -115,7 +115,7 @@ type componentVolumeContext struct {
 }
 
 func (c *componentVolumeContext) ComponentID() string {
-	return fmt.Sprintf("%s/%s:%s", c.cluster.GetNamespace(), c.cluster.GetName(), c.status.GetMemberType())
+	return fmt.Sprintf("%s/%s:%s", c.cluster.GetNamespace(), c.cluster.GetName(), c.status.MemberType())
 }
 
 type pvcResizer struct {
@@ -129,7 +129,7 @@ func NewPVCResizer(deps *controller.Dependencies) PVCResizerInterface {
 }
 
 func (p *pvcResizer) Sync(tc *v1alpha1.TidbCluster) error {
-	components := v1alpha1.ComponentStatusFromTC(tc)
+	components := v1alpha1.AllComponentStatusFromTC(tc)
 	errs := []error{}
 
 	for _, comp := range components {
@@ -152,7 +152,7 @@ func (p *pvcResizer) Sync(tc *v1alpha1.TidbCluster) error {
 }
 
 func (p *pvcResizer) SyncDM(dc *v1alpha1.DMCluster) error {
-	components := v1alpha1.ComponentStatusFromDC(dc)
+	components := v1alpha1.AllComponentStatusFromDC(dc)
 	errs := []error{}
 
 	for _, comp := range components {
@@ -175,7 +175,7 @@ func (p *pvcResizer) SyncDM(dc *v1alpha1.DMCluster) error {
 }
 
 func (p *pvcResizer) buildContextForTC(tc *v1alpha1.TidbCluster, status v1alpha1.ComponentStatus) (*componentVolumeContext, error) {
-	comp := status.GetMemberType()
+	comp := status.MemberType()
 
 	ctx := &componentVolumeContext{
 		cluster:               tc,
@@ -259,7 +259,7 @@ func (p *pvcResizer) buildContextForTC(tc *v1alpha1.TidbCluster, status v1alpha1
 }
 
 func (p *pvcResizer) buildContextForDM(dc *v1alpha1.DMCluster, status v1alpha1.ComponentStatus) (*componentVolumeContext, error) {
-	comp := status.GetMemberType()
+	comp := status.MemberType()
 
 	ctx := &componentVolumeContext{
 		cluster:               dc,
@@ -572,7 +572,7 @@ func (p *pvcResizer) resizeVolumesForPod(ctx *componentVolumeContext, pod *corev
 func (p *pvcResizer) beforeResizeForPod(ctx *componentVolumeContext, resizePod *corev1.Pod, volumes []*volume) error {
 	logPrefix := fmt.Sprintf("before resizing volumes of Pod %s/%s for %q", resizePod.Namespace, resizePod.Name, ctx.ComponentID())
 
-	switch ctx.status.GetMemberType() {
+	switch ctx.status.MemberType() {
 	case v1alpha1.TiKVMemberType:
 		tc, ok := ctx.cluster.(*v1alpha1.TidbCluster)
 		if !ok {
@@ -638,7 +638,7 @@ func (p *pvcResizer) beginResize(ctx *componentVolumeContext) error {
 }
 
 func (p *pvcResizer) endResize(ctx *componentVolumeContext) error {
-	switch ctx.status.GetMemberType() {
+	switch ctx.status.MemberType() {
 	case v1alpha1.TiKVMemberType:
 		// ensure all eviction annotations are removed
 		for _, podVolume := range ctx.actualPodVolumes {
