+++ pkg/manager/member/pvc_resizer_test.go
@@ -153,7 +153,7 @@ func TestResizeVolumes(t *testing.T) {
 			},
 			expect: func(g *WithT, resizer *pvcResizer, ctx *componentVolumeContext, err error) {
 				g.Expect(err).Should(Succeed())
-				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.GetMemberType())).Should(BeFalse())
+				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.MemberType())).Should(BeFalse())
 			},
 		},
 		"begin resizing": {
@@ -195,7 +195,7 @@ func TestResizeVolumes(t *testing.T) {
 			expect: func(g *WithT, resizer *pvcResizer, ctx *componentVolumeContext, err error) {
 				g.Expect(err).Should(HaveOccurred())
 				g.Expect(err.Error()).Should(ContainSubstring("set condition before resizing volumes"))
-				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.GetMemberType())).Should(BeTrue())
+				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.MemberType())).Should(BeTrue())
 			},
 		},
 		"need to resize some volumes": {
@@ -243,7 +243,7 @@ func TestResizeVolumes(t *testing.T) {
 			},
 			expect: func(g *WithT, resizer *pvcResizer, ctx *componentVolumeContext, err error) {
 				g.Expect(err).Should(Succeed())
-				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.GetMemberType())).Should(BeTrue())
+				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.MemberType())).Should(BeTrue())
 
 				cli := resizer.deps.KubeClientset.CoreV1().PersistentVolumeClaims(ctx.cluster.GetNamespace())
 
@@ -305,7 +305,7 @@ func TestResizeVolumes(t *testing.T) {
 			},
 			expect: func(g *WithT, resizer *pvcResizer, ctx *componentVolumeContext, err error) {
 				g.Expect(err).Should(Succeed())
-				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.GetMemberType())).Should(BeTrue())
+				g.Expect(ctx.cluster.(*v1alpha1.TidbCluster).IsComponentVolumeResizing(ctx.status.MemberType())).Should(BeTrue())
 
 				cli := resizer.deps.KubeClientset.CoreV1().PersistentVolumeClaims(ctx.cluster.GetNamespace())
